+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- ![STAT] Improve Canova-Hansen tests for trading days and for seasonality (add options and change R interface)
+
 ## [3.2.4] - 2024-07-11
 
 ### Fixed